+++ ui/src/api/types.ts
@@ -116,6 +116,23 @@ export interface ReleaseSpec {
   channels?: ReleaseChannel[];
 }
 
+export interface HelmChart {
+  name: string;
+  version?: string;
+  path?: string;
+  imagePaths?: string[];
+}
+
+export interface HelmRepository {
+  name: string;
+  url: string;
+  charts: HelmChart[];
+}
+
+export interface HelmSpec {
+  repositories: HelmRepository[];
+}
+
 export interface BlockedImagesSpec {
   blockedImages: string[];
 }